Split schema and sitemap helpers
