+++ frontend/lib/sitemapData.ts
@@ -1,334 +1,23 @@
-import fs from 'node:fs';
-import path from 'node:path';
-import { spawnSync } from 'node:child_process';
 import { defaultLocale, type AppLocale } from '@/i18n/locales';
 import modelRoster from '@/config/model-roster.json';
-import { INDEXED_MARKETING_EXAMPLE_CANONICAL_SLUGS } from '@/config/model-families';
-import {
-  BLOG_ENTRIES,
-  BLOG_SLUG_MAP,
-  CONTENT_ROOT,
-  localizePathFromEnglish,
-} from '@/lib/i18n/paths';
-import { getContentEntries } from '@/lib/content/markdown';
-import { SITEMAP_MANUAL_TIMESTAMPS } from '@/config/sitemap-timestamps';
-import compareConfig from '@/config/compare-config.json';
-import { getHubComparisonSlugsForSitemap } from '@/lib/compare-hub/data';
+import { localizePathFromEnglish } from '@/lib/i18n/paths';
 import { HREFLANG_VARIANTS } from '@/lib/seo/alternateLocales';
-import { SITE_ORIGIN } from '@/lib/siteOrigin';
-import { listEligibleSeoWatchVideos } from '@/server/video-seo';
+import { getCanonicalPathEntries } from './sitemap/route-discovery';
+import {
+  formatLastModified,
+  getLatestEntryDate,
+  getManualSitemapLastModified,
+  getModelLastModified,
+  getModelsSitemapLastModified,
+  getSitemapFileName,
+  getVideoSitemapLastModified,
+} from './sitemap/lastmod';
+import { hasModelLocale } from './sitemap/model-locales';
+import { LOCALES, LOCALE_SITEMAP_PATHS, type SitemapEntry } from './sitemap/types';
+import { buildAbsoluteUrl, escapeXml } from './sitemap/xml';
 
-export type SitemapEntry = {
-  url: string;
-  lastModified?: string;
-  englishPath: string;
-  locales?: AppLocale[];
-  disableAlternates?: boolean;
-};
-
-const SITE_URL = SITE_ORIGIN.replace(/\/+$/, '');
-
-const LOCALE_SITEMAP_PATHS: Record<AppLocale, string> = {
-  en: '/sitemap-en.xml',
-  fr: '/sitemap-fr.xml',
-  es: '/sitemap-es.xml',
-};
-
-const LOCALES: AppLocale[] = ['en', 'fr', 'es'];
-const MODEL_CONTENT_ROOT = path.join(CONTENT_ROOT, 'models');
-
-function findAppPathsManifestPath(): string | null {
-  let currentDir = __dirname;
-  for (let depth = 0; depth < 10; depth += 1) {
-    const manifestPath = path.join(currentDir, 'app-paths-manifest.json');
-    if (fs.existsSync(manifestPath)) {
-      return manifestPath;
-    }
-    const parent = path.dirname(currentDir);
-    if (parent === currentDir) {
-      break;
-    }
-    currentDir = parent;
-  }
-
-  const fallbacks = [
-    path.join(process.cwd(), '.next', 'server', 'app-paths-manifest.json'),
-    path.join(process.cwd(), 'frontend', '.next', 'server', 'app-paths-manifest.json'),
-  ];
-
-  for (const candidate of fallbacks) {
-    if (fs.existsSync(candidate)) {
-      return candidate;
-    }
-  }
-
-  return null;
-}
-
-function resolveSourceAppRoot(): string | null {
-  const candidates = [
-    path.join(process.cwd(), 'app'),
-    path.join(process.cwd(), 'frontend', 'app'),
-    path.join(process.cwd(), '..', 'app'),
-    path.join(process.cwd(), '..', 'frontend', 'app'),
-    path.resolve(__dirname, '..', '..', '..', '..', 'app'),
-    path.resolve(__dirname, '..', '..', '..', '..', '..', 'app'),
-  ];
-
-  const visited = new Set<string>();
-  for (const candidate of candidates) {
-    if (!candidate || visited.has(candidate)) {
-      continue;
-    }
-    visited.add(candidate);
-    if (fs.existsSync(candidate)) {
-      return candidate;
-    }
-  }
-
-  return null;
-}
-
-const APP_PATHS_MANIFEST_PATH = findAppPathsManifestPath();
-const SOURCE_APP_ROOT = resolveSourceAppRoot();
-const LOCALIZED_SOURCE_APP_ROOT = SOURCE_APP_ROOT ? path.join(SOURCE_APP_ROOT, '(localized)', '[locale]') : null;
-const PAGE_FILE_PATTERN = /^page\.(?:mdx|tsx?|ts|jsx?|js)$/i;
-const IGNORED_ROUTE_TEMPLATES = new Set(['/404', '/video/[videoId]', '/v/[videoId]', '/legal/cookies']);
-const MANUAL_ROUTE_DATES = new Map<string, string>(Object.entries(SITEMAP_MANUAL_TIMESTAMPS.routes ?? {}));
-const MANUAL_SITEMAP_DATES = new Map<string, string>(Object.entries(SITEMAP_MANUAL_TIMESTAMPS.sitemaps ?? {}));
-const GIT_LASTMOD_CACHE = new Map<string, string | null>();
-let cachedAppPathsManifest: Record<string, string> | null = null;
-const USE_MTIME_FALLBACK =
-  process.env.SITEMAP_USE_MTIME_FALLBACK === 'true' ||
-  (process.env.SITEMAP_USE_MTIME_FALLBACK !== 'false' && process.env.NODE_ENV !== 'production');
-const FALLBACK_LASTMOD = formatLastModified(process.env.SITEMAP_FALLBACK_LASTMOD);
-
-type CanonicalPathEntry = {
-  englishPath: string;
-  lastModified?: string;
-  locales?: AppLocale[];
-  disableAlternates?: boolean;
-};
-
-type RouteTemplate = {
-  template: string;
-  isDynamic: boolean;
-  sourceFile?: string;
-};
-
-type DynamicRouteGenerator = () => Promise<CanonicalPathEntry[]>;
-
-const parsedTolerance = Number(process.env.SITEMAP_LOCALE_TOLERANCE ?? '3');
-const LOCALE_MISMATCH_TOLERANCE = Number.isFinite(parsedTolerance) && parsedTolerance >= 0 ? parsedTolerance : 3;
-const FAIL_ON_LOCALE_MISMATCH = process.env.SITEMAP_LOCALE_FAIL_ON_MISMATCH === 'true';
-
-const BASE_EXTRA_CANONICAL_PATHS: CanonicalPathEntry[] = [
-  { englishPath: '/legal/cookies-list', locales: ['en', 'fr', 'es'] },
-  { englishPath: '/legal/mentions', locales: ['en', 'fr', 'es'] },
-  { englishPath: '/legal/subprocessors', locales: ['en', 'fr', 'es'] },
-];
-
-function loadAppPathsManifest(): Record<string, string> {
-  if (cachedAppPathsManifest) {
-    return cachedAppPathsManifest;
-  }
-  const manifestPath = APP_PATHS_MANIFEST_PATH;
-  if (!manifestPath) {
-    cachedAppPathsManifest = {};
-    return cachedAppPathsManifest;
-  }
-  try {
-    const raw = fs.readFileSync(manifestPath, 'utf8');
-    cachedAppPathsManifest = JSON.parse(raw) as Record<string, string>;
-  } catch (error) {
-    console.warn('[sitemap] Failed to load app-paths manifest', error);
-    cachedAppPathsManifest = {};
-  }
-  return cachedAppPathsManifest;
-}
-
-function normalizeManifestRoute(route: string): string | null {
-  const prefix = '/(localized)/[locale]';
-  if (!route.startsWith(prefix)) {
-    return null;
-  }
-  const segments = route
-    .slice(prefix.length)
-    .split('/')
-    .filter((segment) => Boolean(segment));
-  const filtered = segments.filter(
-    (segment) =>
-      segment !== 'page' &&
-      segment !== 'route' &&
-      segment !== 'default' &&
-      !isRouteGroupSegment(segment) &&
-      !isParallelRouteSegment(segment)
-  );
-  if (filtered.length === 0) {
-    return '/';
-  }
-  return `/${filtered.join('/')}`;
-}
-
-function findSourceFileForManifestEntry(relativeEntry: string): string | undefined {
-  if (!SOURCE_APP_ROOT) {
-    return undefined;
-  }
-  const trimmed = relativeEntry.startsWith('app/') ? relativeEntry.slice(4) : relativeEntry;
-  const directoryPath = path.dirname(trimmed);
-  const baseName = path.basename(trimmed, path.extname(trimmed));
-  const candidateBases = ['page', 'route', 'default', baseName];
-  const extensions = ['.tsx', '.ts', '.jsx', '.js', '.mdx', '.md'];
-
-  for (const base of candidateBases) {
-    if (!base) continue;
-    for (const ext of extensions) {
-      const candidate = path.join(SOURCE_APP_ROOT, directoryPath, `${base}${ext}`);
-      if (fs.existsSync(candidate)) {
-        return candidate;
-      }
-    }
-  }
-  return undefined;
-}
-
-export const escapeXml = (value: string) =>
-  value
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;')
-    .replace(/'/g, '&apos;');
-
-function buildAbsoluteUrl(pathname: string): string {
-  const normalized = pathname === '/' ? '/' : pathname.startsWith('/') ? pathname : `/${pathname}`;
-  if (normalized === '/' || normalized === '') {
-    return SITE_URL;
-  }
-  return `${SITE_URL}${normalized}`;
-}
-
-function getSitemapFileName(pathname: string): string {
-  const trimmed = pathname.replace(/^\/+/, '');
-  return trimmed || 'sitemap.xml';
-}
-
-function hasModelLocale(slug: string, locale: AppLocale): boolean {
-  if (locale === 'en') {
-    return true;
-  }
-  const candidate = path.join(MODEL_CONTENT_ROOT, locale, `${slug}.json`);
-  return fs.existsSync(candidate);
-}
-
-function hasBlogLocale(canonicalSlug: string, locale: AppLocale): boolean {
-  if (locale === 'en') {
-    return true;
-  }
-  const mapping = BLOG_SLUG_MAP.get(canonicalSlug);
-  return Boolean(mapping?.[locale]);
-}
-
-function formatLastModified(value?: string): string | undefined {
-  if (!value) {
-    return undefined;
-  }
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) {
-    return undefined;
-  }
-  const isoDate = date.toISOString().slice(0, 10);
-  const today = new Date().toISOString().slice(0, 10);
-  return isoDate > today ? today : isoDate;
-}
-
-function getManualRouteLastModified(englishPath: string): string | undefined {
-  return formatLastModified(MANUAL_ROUTE_DATES.get(englishPath));
-}
-
-function getManualSitemapLastModified(name: string): string | undefined {
-  return formatLastModified(MANUAL_SITEMAP_DATES.get(name));
-}
-
-function getGitLastModified(sourceFile?: string): string | undefined {
-  if (!sourceFile || !fs.existsSync(sourceFile)) {
-    return undefined;
-  }
-  const cached = GIT_LASTMOD_CACHE.get(sourceFile);
-  if (cached !== undefined) {
-    return cached ?? undefined;
-  }
-  const relative = path.relative(process.cwd(), sourceFile);
-  const result = spawnSync('git', ['log', '-1', '--pretty=format:%cs', '--', relative], {
-    encoding: 'utf8',
-  });
-  if (!result.error && result.status === 0) {
-    const formatted = formatLastModified(result.stdout.trim());
-    GIT_LASTMOD_CACHE.set(sourceFile, formatted ?? null);
-    return formatted;
-  }
-  if (FALLBACK_LASTMOD) {
-    GIT_LASTMOD_CACHE.set(sourceFile, FALLBACK_LASTMOD);
-    return FALLBACK_LASTMOD;
-  }
-  if (!USE_MTIME_FALLBACK) {
-    GIT_LASTMOD_CACHE.set(sourceFile, null);
-    return undefined;
-  }
-  try {
-    const stats = fs.statSync(sourceFile);
-    const formatted = formatLastModified(stats.mtime.toISOString());
-    GIT_LASTMOD_CACHE.set(sourceFile, formatted ?? null);
-    return formatted;
-  } catch {
-    GIT_LASTMOD_CACHE.set(sourceFile, null);
-    return undefined;
-  }
-}
-
-function getRouteLastModified(englishPath: string, sourceFile?: string): string | undefined {
-  return getManualRouteLastModified(englishPath) ?? getGitLastModified(sourceFile);
-}
-
-function getModelLastModified(slug: string): string | undefined {
-  if (!slug) {
-    return undefined;
-  }
-  const englishPath = `/models/${slug}`;
-  const manual = getManualRouteLastModified(englishPath);
-  if (manual) {
-    return manual;
-  }
-  const candidate = path.join(MODEL_CONTENT_ROOT, 'en', `${slug}.json`);
-  return getGitLastModified(candidate);
-}
-
-function getLatestEntryDate(entries: { lastModified?: string }[]): string | undefined {
-  let latest: string | undefined;
-  entries.forEach((entry) => {
-    if (!entry.lastModified) {
-      return;
-    }
-    if (!latest || entry.lastModified > latest) {
-      latest = entry.lastModified;
-    }
-  });
-  return latest;
-}
-
-async function getVideoSitemapLastModified(): Promise<string | undefined> {
-  const manual = getManualSitemapLastModified(getSitemapFileName('/sitemap-video.xml'));
-  if (manual) {
-    return manual;
-  }
-
-  const watchVideos = await listEligibleSeoWatchVideos();
-  return getLatestEntryDate(
-    watchVideos.map(({ entry, video }) => ({
-      lastModified: formatLastModified(entry.publishedAt || video.createdAt),
-    }))
-  );
-}
+export type { SitemapEntry };
+export { escapeXml, LOCALES, LOCALE_SITEMAP_PATHS };
 
 export async function getLocaleSitemapEntries(locale: AppLocale): Promise<SitemapEntry[]> {
   const seen = new Set<string>();
@@ -491,314 +180,4 @@ export async function buildModelsSitemapXml(): Promise<string> {
   return `<?xml version="1.0" encoding="UTF-8"?>\n<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml">\n${entries.join(
     '\n'
   )}\n</urlset>`;
-}
-
-function getModelsSitemapLastModified(): string | undefined {
-  let latest: string | undefined;
-  modelRoster.forEach((model) => {
-    if (!model?.modelSlug || model?.surfaces?.modelPage?.includeInSitemap === false) {
-      return;
-    }
-    const lastModified = getModelLastModified(model.modelSlug);
-    if (!lastModified) {
-      return;
-    }
-    if (!latest || lastModified > latest) {
-      latest = lastModified;
-    }
-  });
-  return latest;
-}
-
-export { LOCALES, LOCALE_SITEMAP_PATHS };
-
-async function getCanonicalPathEntries(): Promise<CanonicalPathEntry[]> {
-  return resolveCanonicalPathEntries();
-}
-
-async function resolveCanonicalPathEntries(): Promise<CanonicalPathEntry[]> {
-  const templates = discoverLocalizedRouteTemplates();
-  const entries: CanonicalPathEntry[] = [];
-  const seen = new Set<string>();
-  const dynamicTemplates = new Set<string>();
-
-  templates.forEach((template) => {
-    if (template.template === '/ai-video-engines/best-for' && !(compareConfig.bestForPages ?? []).length) {
-      return;
-    }
-    if (template.isDynamic) {
-      dynamicTemplates.add(template.template);
-      return;
-    }
-    const normalizedTemplate = normalizeCompareEnglishPath(template.template);
-    if (seen.has(normalizedTemplate)) {
-      return;
-    }
-    seen.add(normalizedTemplate);
-    const lastModified = getRouteLastModified(normalizedTemplate, template.sourceFile);
-    entries.push({ englishPath: normalizedTemplate, lastModified });
-  });
-
-  for (const template of Array.from(dynamicTemplates)) {
-    const generator = DYNAMIC_ROUTE_GENERATORS[template];
-    if (!generator) {
-      console.warn(`[sitemap] No generator registered for dynamic route ${template}, skipping.`);
-      continue;
-    }
-    let generated: CanonicalPathEntry[] = [];
-    try {
-      generated = await generator();
-    } catch (error) {
-      console.error(`[sitemap] Failed to build entries for ${template}`, error);
-      continue;
-    }
-    generated.forEach((entry) => {
-      if (!entry?.englishPath) {
-        return;
-      }
-      const normalizedPath = normalizeCompareEnglishPath(entry.englishPath);
-      if (seen.has(normalizedPath)) {
-        return;
-      }
-      seen.add(normalizedPath);
-      entries.push({ ...entry, englishPath: normalizedPath, ...(entry.locales ? { locales: entry.locales } : {}) });
-    });
-  }
-
-  const extraCanonicalPaths: CanonicalPathEntry[] = [
-    ...BASE_EXTRA_CANONICAL_PATHS,
-    ...(await listEligibleSeoWatchVideos()).map(({ entry, video }) => ({
-      englishPath: `/video/${entry.id}`,
-      lastModified: formatLastModified(entry.publishedAt || video.createdAt),
-      locales: ['en'] as AppLocale[],
-      disableAlternates: true,
-    })),
-  ];
-
-  extraCanonicalPaths.forEach((extra) => {
-    if (!extra?.englishPath) {
-      return;
-    }
-    const normalizedPath = normalizeCompareEnglishPath(extra.englishPath);
-    if (seen.has(normalizedPath)) {
-      return;
-    }
-    seen.add(normalizedPath);
-    entries.push({
-      ...extra,
-      englishPath: normalizedPath,
-      ...(extra.locales ? { locales: extra.locales } : {}),
-      ...(typeof extra.disableAlternates === 'boolean' ? { disableAlternates: extra.disableAlternates } : {}),
-      lastModified: extra.lastModified ?? getRouteLastModified(normalizedPath),
-    });
-  });
-
-  entries.sort((a, b) => comparePaths(a.englishPath, b.englishPath));
-  validateLocaleCounts(entries);
-  return entries;
-}
-
-function discoverLocalizedRouteTemplates(): RouteTemplate[] {
-  const manifestTemplates = discoverTemplatesFromManifest();
-  if (manifestTemplates.length > 0) {
-    return manifestTemplates;
-  }
-  return discoverTemplatesFromFilesystem();
-}
-
-function discoverTemplatesFromManifest(): RouteTemplate[] {
-  const manifest = loadAppPathsManifest();
-  const entries = Object.entries(manifest);
-  if (!entries.length) {
-    return [];
-  }
-
-  const map = new Map<string, RouteTemplate>();
-
-  entries.forEach(([appRoute, relativeFile]) => {
-    const normalized = normalizeManifestRoute(appRoute);
-    if (!normalized || IGNORED_ROUTE_TEMPLATES.has(normalized)) {
-      return;
-    }
-    const sourceFile = findSourceFileForManifestEntry(relativeFile);
-    const existing = map.get(normalized);
-    if (existing) {
-      if (!existing.sourceFile && sourceFile) {
-        existing.sourceFile = sourceFile;
-      }
-      return;
-    }
-    map.set(normalized, { template: normalized, isDynamic: normalized.includes('['), sourceFile });
-  });
-
-  return Array.from(map.values()).sort((a, b) => comparePaths(a.template, b.template));
-}
-
-function discoverTemplatesFromFilesystem(): RouteTemplate[] {
-  if (!LOCALIZED_SOURCE_APP_ROOT || !fs.existsSync(LOCALIZED_SOURCE_APP_ROOT)) {
-    return [];
-  }
-
-  const stack = [LOCALIZED_SOURCE_APP_ROOT];
-  const map = new Map<string, RouteTemplate>();
-
-  while (stack.length > 0) {
-    const current = stack.pop() as string;
-    const dirents = safeReadDir(current);
-    dirents.forEach((entry) => {
-      const fullPath = path.join(current, entry.name);
-      if (entry.isDirectory()) {
-        stack.push(fullPath);
-        return;
-      }
-      if (!entry.isFile() || !PAGE_FILE_PATTERN.test(entry.name)) {
-        return;
-      }
-      const normalized = buildEnglishPathFromDir(path.dirname(fullPath));
-      if (!normalized || IGNORED_ROUTE_TEMPLATES.has(normalized)) {
-        return;
-      }
-      map.set(normalized, { template: normalized, isDynamic: normalized.includes('['), sourceFile: fullPath });
-    });
-  }
-
-  return Array.from(map.values()).sort((a, b) => comparePaths(a.template, b.template));
-}
-
-function buildEnglishPathFromDir(directory: string): string | null {
-  if (!LOCALIZED_SOURCE_APP_ROOT) {
-    return null;
-  }
-  const relative = path.relative(LOCALIZED_SOURCE_APP_ROOT, directory);
-  if (relative.startsWith('..')) {
-    return null;
-  }
-  const rawSegments = relative.split(path.sep).filter(Boolean);
-  const segments = rawSegments.filter((segment) => !isRouteGroupSegment(segment) && !isParallelRouteSegment(segment));
-  if (!segments.length) {
-    return '/';
-  }
-  return `/${segments.join('/')}`;
-}
-
-function isRouteGroupSegment(segment: string): boolean {
-  return segment.startsWith('(') && segment.endsWith(')');
-}
-
-function isParallelRouteSegment(segment: string): boolean {
-  return segment.startsWith('@');
-}
-
-function comparePaths(a: string, b: string): number {
-  if (a === b) {
-    return 0;
-  }
-  if (a === '/') {
-    return -1;
-  }
-  if (b === '/') {
-    return 1;
-  }
-  return a.localeCompare(b);
-}
-
-function safeReadDir(directory: string): fs.Dirent[] {
-  try {
-    return fs.readdirSync(directory, { withFileTypes: true });
-  } catch {
-    return [];
-  }
-}
-
-function canonicalizeCompareSlug(slug: string): string {
-  const parts = slug.split('-vs-');
-  if (parts.length !== 2) return slug;
-  const [left, right] = parts;
-  return [left, right].sort().join('-vs-');
-}
-
-function normalizeCompareEnglishPath(pathname: string): string {
-  if (!pathname.startsWith('/ai-video-engines/')) {
-    return pathname;
-  }
-  const trimmed = pathname.split('?')[0]?.replace(/\/+$/, '') || '';
-  if (!trimmed) {
-    return pathname;
-  }
-  const slug = trimmed.slice('/ai-video-engines/'.length);
-  if (!slug || slug.includes('/')) {
-    return pathname;
-  }
-  if (!slug.includes('-vs-')) {
-    return pathname;
-  }
-  const canonicalSlug = canonicalizeCompareSlug(slug);
-  return `/ai-video-engines/${canonicalSlug}`;
-}
-
-const DYNAMIC_ROUTE_GENERATORS: Record<string, DynamicRouteGenerator> = {
-  '/blog/[slug]': async () =>
-    BLOG_ENTRIES.map((entry) => ({
-      englishPath: `/blog/${entry.canonicalSlug}`,
-      lastModified: entry.lastModified,
-      locales: LOCALES.filter((locale) => hasBlogLocale(entry.canonicalSlug, locale)),
-    })),
-  '/docs/[slug]': async () => {
-    const docs = await getContentEntries('content/docs');
-    return docs.map((doc) => ({
-      englishPath: `/docs/${doc.slug}`,
-      lastModified: formatLastModified(doc.updatedAt) ?? getGitLastModified(doc.sourcePath),
-    }));
-  },
-  '/examples/[model]': async () =>
-    INDEXED_MARKETING_EXAMPLE_CANONICAL_SLUGS.map((model) => ({
-      englishPath: `/examples/${model}`,
-    })),
-  '/models/[slug]': async () =>
-    modelRoster
-      .filter((model) => Boolean(model?.modelSlug) && model?.surfaces?.modelPage?.includeInSitemap !== false)
-      .map((model) => ({
-        englishPath: `/models/${model.modelSlug}`,
-        lastModified: getModelLastModified(model.modelSlug),
-        locales: LOCALES.filter((locale) => hasModelLocale(model.modelSlug, locale)),
-      })),
-  '/ai-video-engines/[slug]': async () =>
-    Array.from(new Set(getHubComparisonSlugsForSitemap().map((slug) => canonicalizeCompareSlug(slug)))).map((slug) => ({
-      englishPath: `/ai-video-engines/${slug}`,
-      locales: LOCALES,
-    })),
-  '/ai-video-engines/best-for/[usecase]': async () =>
-    (compareConfig.bestForPages ?? []).map((entry: { slug: string }) => ({
-      englishPath: `/ai-video-engines/best-for/${entry.slug}`,
-      locales: LOCALES,
-    })),
-};
-
-function validateLocaleCounts(entries: CanonicalPathEntry[]): void {
-  const counts: Record<AppLocale, number> = { en: 0, fr: 0, es: 0 };
-  entries.forEach((entry) => {
-    if (entry.disableAlternates) {
-      return;
-    }
-    const availableLocales = entry.locales ?? LOCALES;
-    availableLocales.forEach((locale) => {
-      counts[locale] = (counts[locale] ?? 0) + 1;
-    });
-  });
-
-  const englishCount = counts.en ?? 0;
-  LOCALES.forEach((locale) => {
-    if (locale === 'en') {
-      return;
-    }
-    const localeCount = counts[locale] ?? 0;
-    const difference = Math.abs(englishCount - localeCount);
-    if (difference > LOCALE_MISMATCH_TOLERANCE) {
-      const warningMessage = `[sitemap] ${locale.toUpperCase()} sitemap has ${localeCount} URLs vs EN ${englishCount} (diff ${difference}).`;
-      console.warn(warningMessage);
-      if (FAIL_ON_LOCALE_MISMATCH) {
-        throw new Error(warningMessage);
-      }
-    }
-  });
 }
